3-beginner-ai-project/requirements.txt
langgraph
langchain
python-dotenv
langchain-openai
pypdf2